Simpler / more idiomatic condition:
* (not (nil? v)) is just v
* (contains? country-ids (:id k)) is just (country-ids (:id k))

The sorted-indicator-map function is the densest piece of code in the whole thing - it already has several nested calls, a for comprehension with destructuring and a :when qualifier, a sort-by with both a key function and a comparator (I had to go look in the docs to see how that worked!) and a sequence of vector pairs being poured into an empty map.

Using the more idiomatic condition reduces the amount of code and the nested of parentheses and is in keeping with the level of sophistication elsewhere in the function.

I'd probably break this function up (but I'm not going to submit a PR for that until I've seen your feedback on changing the condition).